Seems to be working now
