+++ server/src/sim/optimize.py
@@ -1,11 +1,13 @@
 from printv import printv
 from bunch import Bunch as struct
 from copy import deepcopy
+from matplotlib.pylab import show, figure, subplot, plot, axis, xlim, ylim, legend
+from numpy import array, ones, zeros, concatenate
 
 default_startyear = 2000
 default_endyear = 2030
 
-def optimize(D, objectives=None, constraints=None, name="default", timelimit=60, verbose=2):
+def optimize(D, objectives=None, constraints=None, name="default", ntimepm=1, timelimit=60, progressplot=False, verbose=2):
     """
     Allocation optimization code:
         D is the project data structure
@@ -17,12 +19,12 @@ def optimize(D, objectives=None, constraints=None, name="default", timelimit=60,
     Version: 2014dec01 by cliffk
     """
     
+    # Imports
     from model import model
-    from copy import deepcopy
     from ballsd import ballsd
     from getcurrentbudget import getcurrentbudget
     from makemodelpars import makemodelpars
-    from numpy import array
+    from timevarying import timevarying
     printv('Running optimization...', 1, verbose)
     
     # Set options to update year range
@@ -30,11 +32,12 @@ def optimize(D, objectives=None, constraints=None, name="default", timelimit=60,
     startyear = objectives.get("year").get("start") or default_startyear
     endyear = objectives.get("year").get("end") or default_endyear
     D.opt = setoptions(D.opt, startyear=startyear, endyear=endyear)
+    
     # Make sure objectives and constraints exist
-    if not isinstance(objectives, struct): objectives = defaultobjectives(D, verbose=verbose)
+    if not isinstance(objectives, struct):  objectives = defaultobjectives(D, verbose=verbose)
     if not isinstance(constraints, struct): constraints = defaultconstraints(D, verbose=verbose)
 
-    #save the optimization parameters
+    # save the optimization parameters
     new_optimization = struct()
     new_optimization.name = name
     new_optimization.objectives = objectives
@@ -42,14 +45,14 @@ def optimize(D, objectives=None, constraints=None, name="default", timelimit=60,
 
     D = saveoptimization(D, name, objectives, constraints)
 
-    objectives = deepcopy(objectives)
+    objectives  = deepcopy(objectives)
     constraints = deepcopy(constraints)
 
     # Convert weightings from percentage to number
-    if objectives.outcome.inci: objectives.outcome.inciweight = float( objectives.outcome.inciweight ) / 100.0
-    if objectives.outcome.daly: objectives.outcome.dalyweight = float( objectives.outcome.dalyweight ) / 100.0
-    if objectives.outcome.death: objectives.outcome.deathweight = float( objectives.outcome.deathweight ) / 100.0
-    if objectives.outcome.cost: objectives.outcome.costweight = float( objectives.outcome.costweight ) / 100.0
+    if objectives.outcome.inci:  objectives.outcome.inciweight  = float(objectives.outcome.inciweight) / 100.0
+    if objectives.outcome.daly:  objectives.outcome.dalyweight  = float(objectives.outcome.dalyweight) / 100.0
+    if objectives.outcome.death: objectives.outcome.deathweight = float(objectives.outcome.deathweight) / 100.0
+    if objectives.outcome.cost:  objectives.outcome.costweight  = float(objectives.outcome.costweight) / 100.0
 
     for ob in objectives.money.objectives.keys():
         if objectives.money.objectives[ob].use: objectives.money.objectives[ob].by = float(objectives.money.objectives[ob].by) / 100.0
@@ -68,31 +71,123 @@ def optimize(D, objectives=None, constraints=None, name="default", timelimit=60,
     D.A[1].label = 'Optimal'
     origalloc = deepcopy(array(D.A[1].alloc))
     D.A = D.A[:2] # TODO WARNING KLUDGY
-    
 
+    nprogs = len(origalloc)    
     
-    def objectivecalc(alloc):
+    maxiters = 100
+    global niter
+    niter = 0
+            
+    # Preallocate arrays to store optimisation progress
+    alliters = zeros((nprogs, D.opt.npts, maxiters+1))
+    allobjs  = zeros(maxiters+1)
+    
+    totalspend = sum(origalloc) # Temp # TODO -- set this up for variable budgets
+    
+    def objectivecalc(optimparams):
         """ Calculate the objective function """
-        alloc /= sum(alloc)/sum(origalloc)
+
+        thisalloc = timevarying(optimparams, ntimepm=ntimepm, nprogs=nprogs, t=D.opt.tvec, totalspend=totalspend)        
+        
         newD = deepcopy(D)
-        newD, newcov, newnonhivdalysaverted = getcurrentbudget(newD, alloc)
+        newD, newcov, newnonhivdalysaverted = getcurrentbudget(newD, thisalloc)
         newD.M = makemodelpars(newD.P, newD.opt, withwhat='c', verbose=0)
         S = model(newD.G, newD.M, newD.F[0], newD.opt, verbose=0)
-        objective = S.death.sum() # TEMP
         
+        # Obtain value of the objective function
+        objective = 0 # Preallocate objective value 
+        if objectives.what == 'outcome':
+            for ob in ['inci', 'death', 'daly', 'cost']:
+                if objectives.outcome[ob]: objective += S[ob].sum() * objectives.outcome[ob + 'weight'] # TODO -- can we do 'daly' and 'cost' like this too??
+        else: print('Work to do here') # 'money'
+        
+        if progressplot: # Just to test time-varying stuff
+        
+            # Store values for plotting
+            global niter 
+            alliters[:, :, niter] = thisalloc
+            allobjs[niter] = objective
+            niter += 1        
+            
+            # Plot value of objective function
+            figure(num=100)
+            subplot(1,4,1)
+            plot(range(niter), allobjs[range(niter)])
+            ylim(ymin=0)
+            xlim(xmin=0, xmax=maxiters)        
+            
+            # Plot allocations over iterations
+            subplot(1,4,2)
+            xlim(xmin=0, xmax=maxiters)
+            for prog in range(nprogs): plot(range(niter), alliters[prog, 0, range(niter)])
+    
+            if ntimepm > 1:
+                # Plot this allocation over time
+                subplot(1,4,3)
+                xlim(xmin=D.opt.tvec[0], xmax=D.opt.tvec[-1])
+                for prog in range(nprogs): plot(D.opt.tvec, thisalloc[prog, :])  
+                plot(D.opt.tvec, thisalloc.sum(axis=0), color='k', linewidth=3)
+                ylim(ymin=0, ymax=round(totalspend + 100, -2))
+                
+            legplot = 4 if ntimepm > 1 else 3
+    
+            # Plot legend on the right        
+            subplot(1,4,legplot)
+            for prog in range(nprogs): plot(0, prog)
+            legend(D.G.meta.progs.short, loc = 'center left')
+            axis('off')
+            show()     
+            
         return objective
         
         
         
+    ## Set up parameter vector for time-varying optimisation...
+        
+        
+    stepsize = 100000
+    growsize = 0.01
+        
+    # Initiate probabilities of parameters being selected
+    stepsizes = zeros(nprogs * ntimepm)
+    
+    # Easy access initial allocation indices and turn stepsizes into array
+    ai = range(nprogs)
+    gi = range(nprogs,   nprogs*2) if ntimepm >= 2 else []
+    si = range(nprogs*2, nprogs*3) if ntimepm >= 3 else []
+    ii = range(nprogs*3, nprogs*4) if ntimepm >= 4 else []
+    
+    # Turn stepsizes into array
+    stepsizes[ai] = stepsize
+    stepsizes[gi] = growsize if ntimepm > 1 else 0
+    stepsizes[si] = stepsize
+    stepsizes[ii] = growsize # Not sure that growsize is an appropriate starting point
+    
+    # Initial values of time-varying parameters
+    growthrate = zeros(nprogs)   if ntimepm >= 2 else []
+    saturation = origalloc       if ntimepm >= 3 else []
+    inflection = ones(nprogs)*.5 if ntimepm >= 4 else []
+    
+    # Concatonate parameters to be optimised
+    optimparams = concatenate((origalloc, growthrate, saturation, inflection))
+        
+    parammin = concatenate((zeros(nprogs), ones(nprogs)*-1e9))
+        
     # Run the optimization algorithm
-    optalloc, fval, exitflag, output = ballsd(objectivecalc, origalloc, xmin=0*array(origalloc), timelimit=timelimit, verbose=verbose)
+    optparams, fval, exitflag, output = ballsd(objectivecalc, optimparams, xmin=parammin, \
+    MaxIter=maxiters-1, absinitial=stepsizes, timelimit=timelimit, verbose=verbose,)
     
     # Update the model
-    for i,alloc in enumerate([origalloc,optalloc]):
+    for i, params in enumerate([origalloc, optparams]):
+        
+        
+        
+        alloc = timevarying(params, ntimepm=len(params)/nprogs, nprogs=nprogs, t=D.opt.tvec, totalspend=totalspend)            
+        
         D, D.A[i].coverage, D.A[i].nonhivdalysaverted = getcurrentbudget(D, alloc)
         D.M = makemodelpars(D.P, D.opt, withwhat='c', verbose=2)
         D.A[i].S = model(D.G, D.M, D.F[0], D.opt, verbose=verbose)
-        D.A[i].alloc = alloc # Now that it's run, store total program costs
+        D.A[i].alloc = alloc # This is overwriting a vector with a matrix # TODO -- initiate properly in makedatapars
     
     # Calculate results
     from makeresults import makeresults
