+++ server/src/sim/runsimulation.py
@@ -30,21 +30,21 @@ def runsimulation(D, startyear=2000, endyear=2030, verbose=2, makeplot = 1, dosa
     from makeccocs import makeallccocs
     D = makeallccocs(D, verbose=verbose, makeplot = 0) # Do not plot, ever
 
-#    from getcurrentbudget import getcurrentbudget
-#    D = getcurrentbudget(D) # TODO Add verbose
-
-    # Calculate results
-    from makeresults import makeresults
-    D.R = makeresults(D, allsims, D.opt.quantiles, verbose=verbose)
-
-    # Gather plot data
-    from gatherplotdata import gatheruncerdata
-    D.plot.E = gatheruncerdata(D, D.R, verbose=verbose)
-
-    # Save output
-    if dosave:
-        from dataio import savedata
-        savedata(D.G.projectfilename, D, verbose=verbose)
-
-    printv('...done running simulation for project %s.' % D.G.projectfilename, 2, verbose)
+##    from getcurrentbudget import getcurrentbudget
+##    D = getcurrentbudget(D) # TODO Add verbose
+#
+#    # Calculate results
+#    from makeresults import makeresults
+#    D.R = makeresults(D, allsims, D.opt.quantiles, verbose=verbose)
+#
+#    # Gather plot data
+#    from gatherplotdata import gatheruncerdata
+#    D.plot.E = gatheruncerdata(D, D.R, verbose=verbose)
+#
+#    # Save output
+#    if dosave:
+#        from dataio import savedata
+#        savedata(D.G.projectfilename, D, verbose=verbose)
+#
+#    printv('...done running simulation for project %s.' % D.G.projectfilename, 2, verbose)
     return D
